Refactor account actions controller and routes
