--- src/user-accounts-actions/services/user-accounts-actions.service.ts
@@ -1,4 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class UserAccountsActionsService {}
